third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-recursive-expected.txt

Ensure correct behavior of drawImage with ImageBitmaps constructed from ImageBitmaps that are constructed from images (not pinned to memory) and canvases (pinned to memory).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 255
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
